Death
I made death. Also the brown screen that pops up is a death screen that is sized badly for some reason and I didn't do text.

The death scene is useless, I just forgot to remove it before I wrote this and C O M M I T E D. Remove it probably.

And the corpse stuff happens, but it is really big for some reason.

Also a bug: when you jump and hit a wall before hitting the ground, you can't jump again. Yay.

This concludes my speech